change double to float

+++ src/items/Items.java
@@ -6,7 +6,7 @@ import gameservice.Paint;
  * 游戏内所有主要元素的基类
  */
 abstract public class Items implements Paint {
-    double radius;//几何体最小外接圆半径
+    float radius;//几何体最小外接圆半径
     private Location location = new Location(0, 0);
 
     public Location getLocation() {
