+++ src/items/Targets.java
@@ -30,7 +30,7 @@ abstract public class Targets extends Items {
      */
     abstract public boolean interactBalls(Ball ball);
 
-    public double distanceWithBall(Ball ball) {
-        return Math.sqrt(Math.abs(Math.pow(this.getLocation().getX() - ball.getLocation().getX(), 2) + Math.pow(this.getLocation().getY() - ball.getLocation().getY(), 2)));
+    public float distanceWithBall(Ball ball) {
+        return (float)Math.sqrt(Math.abs(Math.pow(this.getLocation().getX() - ball.getLocation().getX(), 2) + Math.pow(this.getLocation().getY() - ball.getLocation().getY(), 2)));
     }
 }
